⚡️ modular code
