+++ components/Inputs/index.tsx
@@ -1,0 +1,19 @@
+import Input from "./Input";
+import TextArea from "./TextArea";
+import Select from "./Select";
+import PincodeSelect from "./PincodeSelect";
+import Vaccine from "./Vaccine";
+import CowinCode from "./CowinCode";
+import Date from "./Date";
+import Time from "./Time";
+
+export {
+  Input,
+  TextArea,
+  Select,
+  PincodeSelect,
+  Vaccine,
+  CowinCode,
+  Date,
+  Time,
+};
